requirements.txt
SPEEDTEST | Speed test module for Python
CS50
flask
flask_session
werkzeug.security